Rename helper package to client
The helper package now contains only client.go.
Rename it to represent it's actual content.
